+++ src/NuGet.Core/NuGet.CommandLine.XPlat/ReportRenderers/Interfaces/IReportRenderer.cs
@@ -10,6 +10,6 @@ namespace NuGet.CommandLine.XPlat
     {
         void AddProblem(string errorText, ProblemType problemType);
         IEnumerable<ReportProblem> GetProblems();
-        void AddProjectReport(ListPackageReportModel reportProject);
+        void Render(ListPackageReportModel reportProject);
     }
 }
